Addressed pauls comments v2
