+++ Sources/DeploymentTargetLocalhostRuntime/LocalhostRuntimeSupport.swift
@@ -65,77 +65,3 @@ public class LocalhostRuntime<Service: WebService>: DeploymentProviderRuntime {
         LocalhostStartupCommand<Service>.self
     }
 }
-
-public struct LocalhostStructureExporterCommand<Service: WebService>: StructureExporter {
-    public static var configuration: CommandConfiguration {
-        CommandConfiguration(
-            commandName: "local",
-            abstract: "Export web service structure - Localhost",
-            discussion: "Exports an Apodini web service structure for the localhost deployment",
-            version: "0.3.0"
-        )
-    }
-    
-    @Argument(help: "The location of the json file")
-    public var filePath: String
-    
-    @Option(help: "The identifier of the deployment provider")
-    public var identifier: String
-    
-    @Option(help: "The port number for the first-launched child process")
-    public var endpointProcessesBasePort: Int
-    
-    public func run() throws {
-        let app = Application()
-
-        app.storage.set(DeploymentStructureExporterStorageKey.self, to: self)
-        try Service.start(app: app, webService: Service())
-    }
-    
-    public func retrieveStructure(
-        _ endpoints: Set<CollectedEndpointInfo>,
-        config: DeploymentConfig,
-        app: Application
-    ) throws -> AnyDeployedSystem {
-        guard let openApiDocument = app.storage.get(OpenAPI.StorageKey.self)?.document else {
-            throw ApodiniDeployRuntimeSupportError(message: "Unable to get OpenAPI document")
-        }
-        
-        var defaultSystem = try self.retrieveDefaultDeployedSystem(endpoints, config: config, app: app)
-        
-        defaultSystem.userInfo = try openApiDocument.encodeToJSON()
-        defaultSystem.nodes = Set(try defaultSystem.nodes.enumerated().map { idx, node in
-            try node.withUserInfo(LocalhostLaunchInfo(port: self.endpointProcessesBasePort + idx))
-        })
-
-        return defaultSystem
-    }
-    
-    public init() {}
-}
-
-public struct LocalhostStartupCommand<Service: WebService>: DeploymentStartupCommand {
-    public static var configuration: CommandConfiguration {
-        CommandConfiguration(
-            commandName: "local",
-            abstract: "Start a web service - Localhost",
-            discussion: "Starts up an Apodini web service for the localhost deployment",
-            version: "0.3.0"
-        )
-    }
-    
-    @Argument(help: "The location of the json containing the system structure")
-    public var filePath: String
-    
-    @Option(help: "The identifier of the deployment node")
-    public var nodeId: String
-    
-    public func run() throws {
-        let app = Application()
-
-        app.storage.set(DeploymentStartUpStorageKey.self, to: self)
-        try Service.start(app: app, webService: Service())
-    }
-    
-    public init() {}
-}
